+++ packages/schema/src/index.ts
@@ -3,6 +3,7 @@ export {
   internalMetadata,
   jsonField,
   source,
+  strictJSON,
   temporalFormat,
   uiHidden,
   virtual,